Half-done game page
Half done game page
Merged v2 pages into main version pages
