+++ react-frontend/src/App.js
@@ -1,11 +1,13 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import { AnimatePresence } from "framer-motion";
+import { ParallaxProvider } from "react-scroll-parallax";
 import Register from "./pages/Register";
 import Home from "./pages/Home";
+import Home2 from "./pages/Home";
 import Profile from "./pages/Profile";
 import BoardUser from "./components/BoardUser";
 import BoardModerator from "./components/BoardModerator";
+import "./App.css";
 import BoardAdmin from "./components/BoardAdmin";
 import NavBar from "./components/NavBar";
 import Game from "./pages/Game";
@@ -21,20 +23,19 @@ import EventBus from "./common/EventBus";
 
 function App() {
   return (
-    <div className="main">
+    <div className="main overflow-hidden">
       <video
         src="https://tgi-bucket.s3.ap-southeast-1.amazonaws.com/bg_vid.mp4"
         type="video/mp4"
         autoPlay
         loop
         muted
+        className="fixed bg-video"
       />
+      <NavBar className="my-nav" />
       <div className="content">
-        <NavBar />
-
-        <div className="line"></div>
-        <div className=" justify-center  pl-10 pr-10 pt-3 ">
-          <AnimatePresence>
+        <div className="">
+          <ParallaxProvider>
             <Routes>
               <Route path="/" element={<Home />} />
               <Route path="/home" element={<Home />} />
@@ -48,11 +49,12 @@ function App() {
               <Route path="/gameover" element={<GameOver />} />
               <Route path="/gamewin" element={<GameWin />} />
             </Routes>
-          </AnimatePresence>
+          </ParallaxProvider>
         </div>
-        <div className=" flex justify-center items-center pl-10 pr-10 pt-3 ">
+        <div className=" flex justify-center items-center">
           <Routes>
-            <Route path="/leaderboard" element={<Leaderboard />} />
+            {/* <Route path="/leaderboard" element={<Leaderboard />} /> */}
+            <Route path="/leaderboard" element={<Game />} />
           </Routes>
         </div>
         {/* <AuthVerify logOut={this.logOut}/> */}
