--- react-frontend/src/App2.css
@@ -1,182 +1,0 @@
-/* Used in Register and Login */
-label {
-  display: block;
-  margin-top: 10px;
-}
-.card-container.card {
-  max-width: 350px !important;
-  padding: 40px 40px;
-}
-.card {
-  background-color: #f7f7f7;
-  padding: 20px 25px 30px;
-  margin: 0 auto 25px;
-  margin-top: 50px;
-  -moz-border-radius: 2px;
-  -webkit-border-radius: 2px;
-  border-radius: 2px;
-  -moz-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-  -webkit-box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-  box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-}
-.profile-img-card {
-  width: 96px;
-  height: 96px;
-  margin: 0 auto 10px;
-  display: block;
-  -moz-border-radius: 50%;
-  -webkit-border-radius: 50%;
-  border-radius: 50%;
-}
-.components-common-assets-LoginPopup {
-  position: relative;
-  width: 60rem;
-  height: 38rem;
-  background: url(assets/popup_bg.png) 50% / contain no-repeat;
-  padding: 100px 25px 30px;
-  margin: 0 auto 25px;
-  margin-top: 5px;
-  max-width: 350px !important;
-}
-/*  */
-
-/* the line under the nav bar*/
-.border-bottom-line {
-  -webkit-box-shadow: 0px 1px 1px #010101;
-  -moz-box-shadow: 0px 1px 1px #040404;
-  box-shadow: 5px 5px 10px #00500793;
-}
-
-.center {
-  margin-left: auto;
-  margin-right: auto;
-  /* border: 2px solid whitesmoke; */
-  padding: 10px;
-}
-
-.main {
-  width: 100%;
-  height: 100vh;
-}
-
-/* Makes sure bg video is always at the back */
-.bg-video {
-  z-index: 0;
-}
-
-.content {
-  position: absolute;
-  width: 100vw;
-  /* width: 100vw; */
-  /* Height = 100% - height of navbar */
-  height: calc(100%-25vh);
-  top: 30vh;
-  z-index: 10;
-  /* To hide scrollable content below navbar */
-  overflow-y: scroll;
-  overflow-x: hidden;
-  /* Hides miniscrollbar that appears when overflow is used */
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
-}
-
-.content::-webkit-scrollbar {
-  display: none;
-}
-
-@media screen and (min-width: 1024px) {
-  .content {
-    height: calc(100%-20vh);
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .content {
-    height: calc(100%-10vh);
-  }
-}
-
-@media screen and (min-width: 1024px) {
-  .content {
-    top: 10vh;
-  }
-}
-
-video {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-/* For GameOver and GameWin */
-/* container for the game panel */
-.q-container {
-  width: 100%;
-  height: 50%;
-}
-
-/* For Google Translate API */
-.goog-logo-link {
-  display: none !important;
-}
-
-.goog-te-gadget {
-  color: transparent !important;
-  align-self: center;
-  text-align: center;
-}
-
-.goog-te-gadget .goog-te-combo {
-  color: black !important;
-}
-
-.goog-te-combo {
-  transform: translateY(50%);
-}
-
-@media screen and (min-width: 1024px) {
-  .goog-te-combo {
-    transform: translate(-35%, 50%);
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  .goog-te-combo {
-    transform: translateY(50%);
-  }
-}
-
-#google_translate_element {
-  width: auto;
-  height: auto;
-}
-#google_translate_element select {
-  background-color: none;
-  color: #383ffa;
-  /* border: none; */
-  border-radius: 3px;
-  /* margin-right: 10px; */
-  /* padding: 6px 8px; */
-}
-.goog-te-banner-frame.skiptranslate {
-  display: none !important;
-}
-body {
-  top: 0px !important;
-}
-#goog-gt-tt {
-  display: none !important;
-  color: blue;
-  top: 0px !important;
-}
-.goog-tooltip {
-  display: none !important;
-}
-.goog-tooltip:hover {
-  display: none !important;
-}
-.goog-text-highlight {
-  background-color: transparent !important;
-  border: none !important;
-  box-shadow: none !important;
-}
