--- react-frontend/src/css/spinner2.css
@@ -1,16 +1,0 @@
-@keyframes spinner {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-.spinner-container {
-  height: 350px;
-}
-
-.music-spinner {
-  animation: spinner 5s linear infinite;
-}
